Removed redundant tests
